CLEANUP: silence a warning when building on sparc
On Solaris/sparc, getpid() returns pid_t which is not an int :

  src/peers.c: In function `peer_io_handler':
  src/peers.c:508: warning: int format, pid_t arg (arg 6)

+++ src/peers.c
@@ -502,7 +502,7 @@ switchstate:
 				                PEER_SESSION_PROTO_NAME " 1.0\n%s\n%s %d\n%s %lu %d\n",
 				                ps->peer->id,
 				                localpeer,
-				                getpid(),
+				                (int)getpid(),
 				                ps->table->table->id,
 				                ps->table->table->type,
 				                (int)ps->table->table->key_size);
